+++ libs/gui-mockup/src/Flowbox/GuiMockup/LineSnap.hs
@@ -14,14 +14,14 @@ import           Linear
 
 
 
-type PointC = V2 Float
-type LeftHandle = V2 Float
-type RightHandle = V2 Float
-type ControlPoint = (V2 Float, V2 Float, V2 Float)
+type PointC = V2 Double
+type LeftHandle = V2 Double
+type RightHandle = V2 Double
+type ControlPoint = (V2 Double, V2 Double, V2 Double)
 
 --main function for gui
 
-guiLineSnap :: [ControlPoint] -> Maybe ControlPoint -> Maybe ControlPoint -> [V2 Float] -> Float -> [ControlPoint]
+guiLineSnap :: [ControlPoint] -> Maybe ControlPoint -> Maybe ControlPoint -> [V2 Double] -> Double -> [ControlPoint]
 guiLineSnap originalCurveControlPoints pointBefore pointAfter strokePoints errorParameter = resultControlPoints
     where
         resultControlPoints = [startControlPoint] ++ midPoints ++ [endControlPoint]
@@ -35,17 +35,17 @@ guiLineSnap originalCurveControlPoints pointBefore pointAfter strokePoints error
         --resultCurve = moveCurveToStroke (controlPointsToBeziers originalCurveControlPoints) strokePoints errorParameter
         resultCurve = moveCurveToStroke' (controlPointsToBeziers originalCurveControlPoints) pointBefore pointAfter strokePoints errorParameter
 
-moveCurveToStroke :: [CubicBezier Float] -> [V2 Float] -> Float -> [CubicBezier Float]
+moveCurveToStroke :: [CubicBezier Double] -> [V2 Double] -> Double -> [CubicBezier Double]
 moveCurveToStroke originalCurve strokePoints errorParameter = V.toList $ optimizeBeziers (V.fromList originalCurve) (V.fromList strokeAproximation)
     where
         strokeAproximation = fitCurve' strokePoints errorParameter Open
 
-moveCurveToStroke' :: [CubicBezier Float] -> Maybe ControlPoint -> Maybe ControlPoint -> [V2 Float] -> Float -> [CubicBezier Float]
+moveCurveToStroke' :: [CubicBezier Double] -> Maybe ControlPoint -> Maybe ControlPoint -> [V2 Double] -> Double -> [CubicBezier Double]
 moveCurveToStroke' originalCurve pointBefore pointAfter strokePoints errorParameter = V.toList $ optimizeBeziers' (V.fromList originalCurve) pointBefore pointAfter (V.fromList strokeAproximation)
     where
         strokeAproximation = fitCurve' strokePoints errorParameter Open
 
-optimizeBeziers :: V.Vector (CubicBezier Float) -> V.Vector (CubicBezier Float) -> V.Vector (CubicBezier Float) -- V.Vector (V2 Float, V2 Float, V2 Float)
+optimizeBeziers :: V.Vector (CubicBezier Double) -> V.Vector (CubicBezier Double) -> V.Vector (CubicBezier Double) -- V.Vector (V2 Double, V2 Double, V2 Double)
 optimizeBeziers original strokeAproximation =
     let originalDistancesNorm = normalizeDistances $ bezierDistances original
         --originalDistancesNormDoubleEnd = V.snoc originalDistancesNorm (V.last originalDistancesNorm)
@@ -54,7 +54,7 @@ optimizeBeziers original strokeAproximation =
         strokeWithDists = vzip (V.tail strokeDistancesNorm) strokeAproximation
     in V.map (flip generateFittingBezier strokeWithDists) originalDistancesNormWithNeighbours
 
-optimizeBeziers' :: V.Vector (CubicBezier Float) -> Maybe ControlPoint -> Maybe ControlPoint -> V.Vector (CubicBezier Float) -> V.Vector (CubicBezier Float) -- V.Vector (V2 Float, V2 Float, V2 Float)
+optimizeBeziers' :: V.Vector (CubicBezier Double) -> Maybe ControlPoint -> Maybe ControlPoint -> V.Vector (CubicBezier Double) -> V.Vector (CubicBezier Double) -- V.Vector (V2 Double, V2 Double, V2 Double)
 optimizeBeziers' original pointBefore pointAfter strokeAproximation =
     let originalDistancesNorm' = normalizeDistances $ bezierDistances original
         originalDistancesNorm = case (pointBefore, pointAfter) of
@@ -68,26 +68,26 @@ optimizeBeziers' original pointBefore pointAfter strokeAproximation =
         strokeWithDists = vzip (V.tail strokeDistancesNorm) strokeAproximation
     in V.map (flip generateFittingBezier strokeWithDists) originalDistancesNormWithNeighbours
 
-adjustStartDistance :: (V.Vector Float) -> (V.Vector Float)
+adjustStartDistance :: (V.Vector Double) -> (V.Vector Double)
 adjustStartDistance distances = normalizeDistances $ V.map (+1/(fromIntegral $ V.length distances)) distances -- because of normalization step this function has to be applied first
 
-adjustEndDistance :: (V.Vector Float) -> (V.Vector Float)
+adjustEndDistance :: (V.Vector Double) -> (V.Vector Double)
 adjustEndDistance distances = V.map (* ((len-1)/len) ) distances
     where
         len = fromIntegral $ V.length distances
 
 -- prototypes
-adjustStartDistance' :: (V.Vector Float) -> (V.Vector Float)
+adjustStartDistance' :: (V.Vector Double) -> (V.Vector Double)
 adjustStartDistance' distances = normalizeDistances $ V.map (+ V.head distances) distances -- because of normalization step this function has to be applied first
 
-adjustEndDistance' :: (V.Vector Float) -> (V.Vector Float)
+adjustEndDistance' :: (V.Vector Double) -> (V.Vector Double)
 adjustEndDistance' distances = V.map (* proportion ) distances
     where
         proportion = 1/(1.0 + V.last distances)
 
 -- end of prototypes
 
-generateFittingBezier :: (Float, Float) -> V.Vector (Float, CubicBezier Float) -> CubicBezier Float
+generateFittingBezier :: (Double, Double) -> V.Vector (Double, CubicBezier Double) -> CubicBezier Double
 generateFittingBezier (startLength, endLength) strokeWithDists = iteration
     where
         iteration = runST $ do
@@ -132,7 +132,7 @@ generateFittingBezier (startLength, endLength) strokeWithDists = iteration
         (endPoint, hi, _) = deCasteljauCubic endRest endPointBezier
         nextBeziers = V.map snd $ V.tail beforeEndStrokeBeziers
 
---beziersToControlPoints :: V.Vector (CubicBezier Float) -> V.Vector (V2 Float, V2 Float, V2 Float)
+--beziersToControlPoints :: V.Vector (CubicBezier Double) -> V.Vector (V2 Double, V2 Double, V2 Double)
 --beziersToControlPoints beziers = --V.snoc (V.map snd $ V.tail $ V.scanl (\(prevhi,_) (CubicBezier point1 ho hi point2) -> (hi,(point1,prevhi,ho))) (0.0, (firstPoint,dummyHandle,firstPointHo)) beziers) lastControlPoint where
     --CubicBezier firstPoint firstPointHo _ _ = V.head beziers
     --dummyHandle = V2 0 0
@@ -140,14 +140,14 @@ generateFittingBezier (startLength, endLength) strokeWithDists = iteration
     --lastControlPoint = (lastPoint, lastPointHi, lastPoint) -- for js purpose. Change third element to dummyHandle
 
 -- Function returns control points where both handles are well defined (first point of first segment and second point of last segment are ignored)
-beziersToFullControlPoints :: [CubicBezier Float] -> [ControlPoint]
+beziersToFullControlPoints :: [CubicBezier Double] -> [ControlPoint]
 beziersToFullControlPoints beziers = map snd $ tail . tail $ scanl (\(prevhi,_) (CubicBezier point1 ho hi point2) -> (hi,(point1,prevhi,ho))) (0.0, (firstPoint,dummyHandle,firstPointHo)) beziers where
     CubicBezier firstPoint firstPointHo _ _ = head beziers
     dummyHandle = V2 0 0
     CubicBezier _ _ lastPointHi lastPoint = last beziers
     --lastControlPoint = (lastPoint, lastPointHi, lastPoint) -- for js purpose. Change third element to dummyHandle
 
-beziersToControlPoints :: [CubicBezier Float] -> [ControlPoint]
+beziersToControlPoints :: [CubicBezier Double] -> [ControlPoint]
 beziersToControlPoints beziers = [startControlPoint] ++ midPoints ++ [endControlPoint]
     where
         midPoints = beziersToFullControlPoints beziers
@@ -156,17 +156,17 @@ beziersToControlPoints beziers = [startControlPoint] ++ midPoints ++ [endControl
         CubicBezier ps ho _ _ = head beziers
         CubicBezier _ _ hi pe = last beziers
 
-controlPointsToBeziers :: [ControlPoint] -> [CubicBezier Float]
+controlPointsToBeziers :: [ControlPoint] -> [CubicBezier Double]
 controlPointsToBeziers controlPoints =
     tail $ tail $ map snd $ scanl funcRR ((0,0), streight1) controlPoints
         where
             funcRR = (\((x,y),bezier) (p,hi,ho) -> ((p,ho), CubicBezier x y hi p))
 
-deCasteljauCubic :: Float -> CubicBezier Float -> (V2 Float, V2 Float, V2 Float)
+deCasteljauCubic :: Double -> CubicBezier Double -> (V2 Double, V2 Double, V2 Double)
 deCasteljauCubic t bezier = (point, handle1, handle2) where
     [point, handle1, handle2] = deCasteljau t [cubicC0 bezier, cubicC1 bezier, cubicC2 bezier, cubicC3 bezier]
 
-deCasteljau :: Float -> [V2 Float] -> [V2 Float]
+deCasteljau :: Double -> [V2 Double] -> [V2 Double]
 deCasteljau t coefs = --trace ("reduced: "++show reduced) $
     case coefs of
         [c1,c2] -> reduced ++ coefs
@@ -176,15 +176,15 @@ deCasteljau t coefs = --trace ("reduced: "++show reduced) $
         lerpP t (V2 x0 y0) (V2 x1 y1) = V2 (lerp t x0 x1) (lerp t y0 y1)
         lerp t a b = t * b + (1 - t) * a
 
-arcLength :: CubicBezier Float -> Float
+arcLength :: CubicBezier Double -> Double
 arcLength curve =
     let points = map (bezierPoint curve) [0.1, 0.2..1] -- fixed 10 segments division
     in fst $ foldl lengthCounter (0, cubicC0 curve) points
 
-lengthCounter :: (Float, V2 Float) -> V2 Float -> (Float, V2 Float)
+lengthCounter :: (Double, V2 Double) -> V2 Double -> (Double, V2 Double)
 lengthCounter (acc, pt1) pt2 = (acc + euclidianDistance pt1 pt2, pt2)
 
-bezierPoint :: CubicBezier Float -> Float -> V2 Float
+bezierPoint :: CubicBezier Double -> Double -> V2 Double
 bezierPoint curve t =
     let V2 x0 y0 = cubicC0 curve
         V2 x1 y1 = cubicC1 curve
@@ -194,23 +194,23 @@ bezierPoint curve t =
             start*(1-t)*(1-t)*(1-t) + 3.0*control1*(1.0-t)*(1.0-t)*t + 3.0*control2*(1.0-t)*t*t + end*t*t*t
     in V2 (formula x0 x1 x2 x3) (formula y0 y1 y2 y3)
 
-resampleBezier :: CubicBezier Float -> V.Vector (V2 Float)
+resampleBezier :: CubicBezier Double -> V.Vector (V2 Double)
 resampleBezier = resampleBezierFragment 0.0 1.0
 
-resampleBezierFragment :: Float -> Float -> CubicBezier Float -> V.Vector (V2 Float)
+resampleBezierFragment :: Double -> Double -> CubicBezier Double -> V.Vector (V2 Double)
 resampleBezierFragment start end bezier = if start==end then V.singleton (bezierPoint bezier start) else V.map (bezierPoint bezier) (V.fromList [start,start+((end-start)/10)..end])
 
-euclidianDistance :: V2 Float -> V2 Float -> Float
+euclidianDistance :: V2 Double -> V2 Double -> Double
 euclidianDistance (V2 x1 y1) (V2 x2 y2)= sqrt ((x1-x2)*(x1-x2) + (y1-y2)*(y1-y2))
 
 --For stroke aproximation. Replaced by LineFit module
 --strokeDistances points = V.map fst $ V.scanl lengthCounter (0, V.head points) (V.tail points)
 
-bezierDistances :: V.Vector (CubicBezier Float) -> V.Vector Float
+bezierDistances :: V.Vector (CubicBezier Double) -> V.Vector Double
 bezierDistances = V.scanl (\acc c -> acc+arcLength c) 0
 --bezierDistances beziers = map arcLength beziers
 
-normalizeDistances :: V.Vector Float -> V.Vector Float
+normalizeDistances :: V.Vector Double -> V.Vector Double
 normalizeDistances distances = V.map (/ V.last distances) distances
 
 vzip :: (Storable a, Storable b) => V.Vector a -> V.Vector b -> V.Vector (a,b)
@@ -218,12 +218,12 @@ vzip = V.zipWith (,)
 
 
 -- helpers for testing with JFiddle
-getPoints :: [String] -> [V2 Float]
+getPoints :: [String] -> [V2 Double]
 getPoints strings = map (Prelude.uncurry V2) (getJsPoints strings) where
-    getJsPoints :: [String] -> [(Float,Float)]
+    getJsPoints :: [String] -> [(Double,Double)]
     getJsPoints = map read
 
-printJsPoints :: [V2 Float] -> String
+printJsPoints :: [V2 Double] -> String
 printJsPoints points = show $ map (\(V2 x y) -> [floor x, floor y]) points
 
 paperPointsToBeziers points =
@@ -235,26 +235,26 @@ funcR = (\((x,y),bezier) (p,hi,ho) -> ((p,p+ho), CubicBezier x y (p+hi) p))
 --process original = printJsPoints . concatMap (\(x,y,z) -> [x,y-x,z-x]) . V.toList . beziersToControlPoints . optimizeBeziers original . paperPointsToBeziers
 
 -- test data
-s111221 :: V.Vector (CubicBezier Float)
+s111221 :: V.Vector (CubicBezier Double)
 s111221 = V.fromList [streight1,streight1,streight1,streight2,streight2,streight1]
 
-streight1 :: CubicBezier Float
+streight1 :: CubicBezier Double
 streight1 = CubicBezier (V2 0 0) (V2 0.2 0) (V2 0.8 0) (V2 1 0)
 
-streight2 :: CubicBezier Float
+streight2 :: CubicBezier Double
 streight2 = CubicBezier (V2 0 0) (V2 0.4 0) (V2 1.6 0) (V2 2 0)
 
-exampleCurve :: CubicBezier Float
+exampleCurve :: CubicBezier Double
 exampleCurve = CubicBezier (V2 0 0) (V2 4 0) (V2 5 1) (V2 5 5)
 
-exampleCurve2 :: CubicBezier Float
+exampleCurve2 :: CubicBezier Double
 exampleCurve2 = CubicBezier (V2 5 5) (V2 5 10) (V2 15 25) (V2 25 25)
 
-exampleCurve3 :: CubicBezier Float
+exampleCurve3 :: CubicBezier Double
 exampleCurve3 = CubicBezier (V2 25 25) (V2 30 25) (V2 30 40) (V2 50 50)
 
-examplePoints :: V.Vector (V2 Float)
+examplePoints :: V.Vector (V2 Double)
 examplePoints = V.fromList [V2 10 0, V2 11 0, V2 12 1, V2 12 5, V2 12 6]
 
-examplePoints2 :: V.Vector (V2 Float)
+examplePoints2 :: V.Vector (V2 Double)
 examplePoints2 = V.fromList [V2 1 0, V2 5 0, V2 5 5, V2 6 5]
